AgroVision-Backend-Python/app/requirements.txt
fastapi
uvicorn[standard]
pillow
tensorflow==2.16.1
numpy
python-multipart
